# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel API",
        "description": "Sayohat veb-sayti uchun API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://toqtarbay.dbc.uz/api",
            "description": "Production Server"
        },
        {
            "url": "http://travel.test/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/about": {
            "get": {
                "tags": [
                    "About"
                ],
                "summary": "Biz haqimizda ma'lumotini olish",
                "operationId": "c9c8a41eb9893e0f2c10063d6ca06abd",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Til kodi (kk, uz, ru, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "uz"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli javob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Sarguzashtlar olami"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Biz haqimizda batafsil..."
                                                },
                                                "images": {
                                                    "description": "Rasmlar ro'yxati (tartiblangan)",
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "image_path": {
                                                                "type": "string",
                                                                "example": "/storage/uploads/gallery-1.jpg"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ma'lumot topilmadi"
                    }
                }
            }
        },
        "/about/banner": {
            "get": {
                "tags": [
                    "About"
                ],
                "summary": "About sahifasi bannerini olish",
                "operationId": "fa236f7b0e07e57aab528e3a8af1a854",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Til kodi (kk, uz, ru, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "uz"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli javob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Biz haqimizda"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "/storage/uploads/banners/about-banner.jpg"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Banner topilmadi"
                    }
                }
            }
        },
...
    },
...
}